refactor(web): split media transfer storage ownership
